inverse_kinematics/ac_network.py
import torch
import torch.nn as nn
import torch.optim as optim



class ActorNetwork(nn.Module):
    def __init__(self, input_size, output_size):
        super(ActorNetwork, self).__init__()
        # Define the actor network architecture
        pass
    def forward(self, x):
        # Implement the forward pass for the actor network
        pass
class CriticNetwork(nn.Module):
    def __init__(self, input_size, output_size):
        super(CriticNetwork, self).__init__()
        # Define the critic network architecture
        pass
    def forward(self, x, a):
        # Implement the forward pass for the critic network
        pass